requirements.txt
Flask==0.10.1
Flask-Migrate==1.4.0
Flask-SQLAlchemy==2.0
Flask-DebugToolbar==0.10.0
flask-servicelayer==0.0.4

celery==3.0.20
evelink==0.7.3
psycopg2==2.6.1
eveimageserver==0.1.2